Move network.py to tools/
